Move back ruby and js classes to its original place
